src/requirements.txt
Flask==1.1.1
nltk==3.4.5
numpy==1.18.0
pandas==0.25.3
pymongo==3.10.1
sklearn==0.0
python-dotenv==0.10.5
dnspython